requirements.txt
Django==4.0.6
mysqlclient
python-dotenv==0.20.0
python-telegram-bot==13.13
requests==2.28.1